finalize first testmail sending

+++ newsletter/create_DB.py
@@ -1,37 +1,44 @@
 import os
 import sqlite3
-
+import hashlib
+import random
 
 # TODO
 # columns to add
 # frequency of sending: 1, 7, 30, threshhold
 # date created
 
-# check I runnung on entorb.net webserver
-if os.path.isdir("/home/entorb/data-web-pages/covid-19"):
-    pathToDb = '/home/entorb/data-web-pages/covid-19/newsletter.db'
-else:
-    pathToDb = 'cache/newsletter.db'
 
-if os.path.isfile(pathToDb):
-    os.remove(pathToDb)
+def checkRunningOnServer() -> bool:
+    if os.path.isdir("/home/entorb/data-web-pages/covid-19"):
+        return True
+    else:
+        return False
 
-con = sqlite3.connect(pathToDb)
-# con = sqlite3.connect(":memory:")
-# con.row_factory = sqlite3.Row  # allows for access via row["name"]
-cur = con.cursor()
+
+def deleteDB():
+    if not checkRunningOnServer() and os.path.isfile(pathToDb):
+        os.remove(pathToDb)
 
 
 def create_table():
     cur.execute("""
-      CREATE TABLE newsletter (email text, activated int, hash text, threshhold int, regions text)
+      CREATE TABLE newsletter (email text, verified int, hash text, threshhold int, regions text)
       """
                 )
 
 
-def test_insert():
-    cur.execute(f"INSERT INTO newsletter(email, activated, hash, threshhold, regions) VALUES (?,?,?,?,?)",
-                ("test@entorb.net", 1, "<hash>", 250,
+def gen_MD5_string(s: str) -> str:
+    m = hashlib.md5()
+    m.update(s.encode('ascii'))
+    return m.hexdigest()
+
+
+def test_insert(email: str):
+    s = email + str(random.random())
+    h = gen_MD5_string(s)
+    cur.execute(f"INSERT INTO newsletter(email, verified, hash, threshhold, regions) VALUES (?,?,?,?,?)",
+                (email, 1, h, 250,
                  "09562,09572,09563,09564,03353,02000,14612"))
     con.commit()
 
@@ -39,12 +46,24 @@ def test_insert():
 def test_select():
     for row in cur.execute("SELECT * FROM newsletter"):
         print(row)
-    print("We now have %s rows in the DB table" %
+    print("%s rows in the DB table" %
           cur.execute("SELECT count(*) FROM newsletter").fetchone()[0])
 
 
+# check I runnung on entorb.net webserver
+if checkRunningOnServer():
+    pathToDb = '/home/entorb/data-web-pages/covid-19/newsletter.db'
+else:
+    pathToDb = 'cache/newsletter.db'
+
+
+deleteDB()
+
+con = sqlite3.connect(pathToDb)
+cur = con.cursor()
+
 create_table()
-test_insert()
+test_insert("test@entorb.net")
 test_select()
